requirements.txt
Django==5.2.10
djangorestframework==3.16.1
django-cors-headers==4.9.0
gunicorn==21.2.0
psycopg2-binary==2.9.11
whitenoise==6.6.0
joblib==1.5.3
numpy==2.2.6
scikit-learn==1.8.0
pandas==2.3.3
python-decouple==3.8
dj-database-url==2.1.0
